+++ slk/common.py
